% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds_pca.R
\name{calculate_mds_pca}
\alias{calculate_mds_pca}
\title{MDS or PCA}
\usage{
calculate_mds_pca(
  se,
  assay = 1,
  method = "pca",
  dist = "euclidean",
  center = TRUE,
  scale = FALSE,
  no.pc = 3
)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay containing
expression data to be used for dimensionality reduction.}

\item{method}{Method to use: "pca" (default), "mds".}

\item{dist}{Distance method to be used for MDS: see method argument in
\code{\link[stats]{dist}} function.}

\item{center}{Logical. Should the data be centered by mean? (default: TRUE).}

\item{scale}{Logical. Should the data be scaled by standard deviation?
(default: FALSE).}

\item{no.pc}{Integer. Number of components that are returned (default: 3).}
}
\value{
list with components scores (data.frame with components of PCA or
MDS analysis) and var.explained (vector with explained variance; only for
PCA).
}
\description{
Performs dimensionality reduction using multi-dimensional scaling (MDS) or
principal component analysis (PCA) on the sample level.
}
\examples{
data("se.gene")

## PCA
res.pca = calculate_mds_pca(se = se.gene,
                            method = "pca")

## MDS
res.mds = calculate_mds_pca(se = se.gene,
                            method = "mds")
}
