% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_batch_effects.R
\name{check_batch_effects}
\alias{check_batch_effects}
\title{Check for batch effects}
\usage{
check_batch_effects(se, assay = 1, col.batch = "batch", filt = NULL,
  nperm = 1000, center = FALSE, scaleY = FALSE, seed = 42)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay containing
expression data to be used for test of batch effects.}

\item{col.batch}{Character or integer. Column in colData() with batch
information (default: batch as generated by the function
\code{\link{define_batches}}.}

\item{filt}{Numeric. Parameter of the function
\code{\link[gPCA]{gPCA.batchdetect}} (default: NULL meaning no variance
filtering).}

\item{nperm}{Integer. Parameter of the function
\code{\link[gPCA]{gPCA.batchdetect}} (default: 1000 permutations for
empirical P value).}

\item{center}{Logical. Parameter of the function
\code{\link[gPCA]{gPCA.batchdetect}} (default: FALSE for uncentered
expression data).}

\item{scaleY}{Logical. Parameter of the function
\code{\link[gPCA]{gPCA.batchdetect}} (default: FALSE for no scaling by the
number of samples in each batch).}

\item{seed}{Integer. Parameter of the function
\code{\link[gPCA]{gPCA.batchdetect}} (default: 42 instead of NULL to make
function reproducible).}
}
\value{
see \code{\link[gPCA]{gPCA.batchdetect}}
}
\description{
Statistically tests for batch effects using the function
\code{\link[gPCA]{gPCA.batchdetect}}
}
