% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sex.R
\name{check_sex}
\alias{check_sex}
\title{Sex check}
\usage{
check_sex(se, assay = 1, gene.column = "symbol", info.sex.genes,
  sex.column = "sex", col = c("red", "blue"), title = NULL,
  ellipse = TRUE, ellipse.type = "norm")
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay containing
expression data to be used for sex check.}

\item{gene.column}{Character. Column in rowData containing gene symbols
(default: "symbol") or "rownames" if rownames should be used.}

\item{info.sex.genes}{data.frame. Information about sex chromosomal genes
(see \code{info.sex.genes}).}

\item{sex.column}{Character. Column in colData containing sex information.}

\item{col}{Character vector. Colours to be used for coloring sex in MDS plot
(default: red and blue).}

\item{title}{Character. Title of the plot. If NULL title will be set to
"MDS (sex specific genes)".}

\item{ellipse}{Logical. Should ellipses around points be drawn? (default:
TRUE).}

\item{ellipse.type}{Character. Type of ellipse as given in
\code{\link[ggpubr]{ggscatter}} (default: "norm").}

\item{genes}{Character vector. Symbols of sex specific genes to be used in
classification (default: genes collected from the literature, see details).}
}
\value{
List with the following components:
\itemize{
\item info: data.frame with information about samples with wrong sex or NULL
\item plot: MDS plot as returned by \code{plot_mds_pca_2d}
}
}
\description{
Predicts sex using k nearest neighbour classification based on expression
of sex specific genes.
}
\examples{
data("se.gene")
data("info.sex.genes")

check_sex(se = se.gene,
          info.sex.genes = info.sex.genes,
          gene.column = "Gene.symbol",
          sex.column = "Sex")
}
