% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_se_objects.R
\name{combine_se_objects}
\alias{combine_se_objects}
\title{Merge SummarizedExperiments objects}
\usage{
combine_se_objects(se.l, info = NULL)
}
\arguments{
\item{se.l}{list of \code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
objects}

\item{info}{Vector of same length as se.l. Contains additional information
about each SE object (e.g. study name) which will be replicated for each
sample in the corresponding SE object and stored in a column named info.}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with all samples but no rowData() information
}
\description{
Combine several SummarizedExperiments objects into a single
SummarizedExperiments object based on common genes and common colnames in
colData.
}
\examples{
library(SummarizedExperiment)
data("se.gene")

## split SE object into two SE objects (for illustration only)
## and store in list
se.l = list(part.1 = se.gene[, 1:10],
            part.2 = se.gene[, 11:26])

se.combined = combine_se_objects(se.l = se.l)
all.equal(dim(se.gene), dim(se.combined))

## add information about parts
se.combined = combine_se_objects(se.l = se.l,
                                 info = c("part.1", "part.2"))
table(se.combined$info)
}
