% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geo_characteristics.R
\name{convert}
\alias{convert}
\title{internal function used by extract_geo_characteristics}
\usage{
convert(col.char)
}
\description{
extracts name and information from single column
converts to numeric if possible
}
\keyword{internal}
