% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_batch_effects.R
\name{define_batches}
\alias{define_batches}
\title{Define batches based on scan date}
\usage{
define_batches(se, col.scan.date = "scan.date", diff.ignore = 1)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{col.scan.date}{Character or Integer. Column in colData() with scan
dates (default: scan.date as generated by the function
\code{\link{extract_scan_date}}.}

\item{diff.ignore}{Numeric. Time difference (in days) defined as negligible
(default: 1).}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with batch information in additional column in colData()
}
\description{
Groups samples into batches based on scan date. Samples run within a short
time interval can be defined to belong to the same batch.
}
\examples{
data("se.gene")
se.gene = define_batches(se = se.gene,
                         col.scan.date = "scan.date")
}
