% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicated_samples.R
\name{detect_duplicated_samples}
\alias{detect_duplicated_samples}
\title{Identification of duplicated samples}
\usage{
detect_duplicated_samples(se, assay = 1, cor.method = "spearman", ...)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay containing
log transformed expression values.}

\item{cor.method}{Character. Correlation coefficient to be used ("pearson",
"spearman" (default) or "kendall").}

\item{...}{Additional arguments passed to the
\code{\link[doppelgangR]{outlierFinder}} function.}
}
\value{
List with the following components:
\itemize{
\item info: data.frame with identifiers of duplicated samples and their
correlation or NULL
\item plot: Histogram as returned by \code{\link[ggpubr]{gghistogram}}
}
}
\description{
Detects duplicated samples with extremely high correlation of expression
values using the function \code{\link[doppelgangR]{outlierFinder}}.
}
\examples{
data("se.gene")

detect_duplicated_samples(se = se.gene)
}
