% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_data.R
\name{extract_detection_pvalue}
\alias{extract_detection_pvalue}
\title{Extract detection P value}
\usage{
extract_detection_pvalue(accession, se)
}
\arguments{
\item{accession}{Character. Identifier of GEO data set (GSEXXX).}

\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with detection P value added as assay with the name detection.pvalue.
}
\description{
Extracts detection P value information from each sample file.
}
\examples{
data("se.probeset")

se.probeset = extract_detection_pvalue(se = se.probeset,
                                       accession = "GSE6710")
}
