% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geo_characteristics.R
\name{extract_geo_characteristics}
\alias{extract_geo_characteristics}
\title{Extract GEO information}
\usage{
extract_geo_characteristics(se, cols = NULL)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{cols}{Character or integer. Name or number of column(s) in colData with
GEO information (optional, will otherwise be extracted assuming column names
contain term 'characteristics').}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with GEO information in additional columns in colData()
}
\description{
Extracts additional sample information from GEO into separate columns of
colData(). Numeric variables are converted from character
to numeric and syntactically valid variable names are used for the new
columns.
}
