% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_data.R
\name{extract_scan_date}
\alias{extract_scan_date}
\title{Extract scan date}
\usage{
extract_scan_date(
  se,
  col.sample.id = "geo_accession",
  temp.dir = tempdir(),
  tryFormats = c("\%Y-\%m-\%d", "\%Y/\%m/\%d", "\%m/\%d/\%Y", "\%m/\%d/\%y",
    "\%m-\%d-\%Y", "\%d-\%b-\%Y")
)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{col.sample.id}{Character or integer vector. Column in colData() with
sample information.}

\item{temp.dir}{Character. Destination directory to downloaded array files.}

\item{tryFormats}{Character vector. Format strings for date to try.}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with scan date added as column scan.date in colData.
}
\description{
Extracts date of scanning from original array files (e.g. CEL for Affymetrix
and idat for Illumina arrays).
}
\examples{
data("se.probeset")

se.probeset = extract_scan_date(se = se.probeset)
}
