% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_data.R
\name{get_geo_data}
\alias{get_geo_data}
\title{Export gene expression data sets from GEO}
\usage{
get_geo_data(accession, temp.dir = tempdir(), platform = NULL,
  detection.pvalue = FALSE, scan.date = FALSE)
}
\arguments{
\item{accession}{Character. Identifier of GEO data set (GSEXXX).}

\item{temp.dir}{Character. Destination directory to downloaded array files.}

\item{platform}{Character. GEO identifier of platform that should be
selected. Only relevant, if study profiled samples on different platforms.}

\item{detection.pvalue}{Logical. Should information about detection P values
be extracted (if available)?}

\item{scan.date}{Logical. Should information about scan dates be extracted
(if raw data files are available)?}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object.
}
\description{
Downloads normalized gene expression data from Gene Expression Omnibus (GEO)
and stores it as a SummarizedExperiment object. Information about detection
P values and dates of scanning can also be extracted.
}
