% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{get_outliers}
\alias{get_outliers}
\title{internal function used by plot_distribution}
\usage{
get_outliers(values, coef = 5)
}
\description{
determines indices of outliers in values
}
\keyword{internal}
