% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds_pca.R
\name{get_outliers_mds_pca_2d}
\alias{get_outliers_mds_pca_2d}
\title{internal function used by plot_mds_pca}
\usage{
get_outliers_mds_pca_2d(scores, dim, factor = 5)
}
\description{
determines outliers in two dimensional plot
}
\keyword{internal}
