% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{Log transformation}
\usage{
log_transform(se, assay = 1, pseudocount = 1)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay containing
expression data to be log transformed.}

\item{pseudocount}{Numeric. Value added to expression values before log
transformation (relevant for count data containing zero values).}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with additional assay called <assay>.log.
}
\description{
Performs log transformation of expression values (log2). Values below zero
will be set to zero and if zero values are available
}
\examples{
data("se.probeset")

se.probeset = log_transform(se = se.probeset)
}
