% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot distribution of expression values for each sample}
\usage{
plot_distribution(
  se,
  assay = 1,
  method = "quantileplot",
  coef = 5,
  label = "expression value",
  title = paste("Distribution of expression", "values in each sample")
)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay to be used for
plotting.}

\item{method}{Method to use for plotting: "quantileplot" (default),
"boxplot", "violinplot".}

\item{coef}{Numeric. Used in outlier definition (median +/- coef * IQR)}

\item{label}{Character. Label of x or y axis (default: "expression value").}

\item{title}{Character. Title of the plot.}
}
\value{
List with the following components:
\itemize{
\item info: data.frame with information about outlier samples or NULL
\item plot: plot as returned by \code{\link[GGally]{ggparcoord}} (quantile
plot), \code{\link[ggplot2]{ggplot}} (boxplot) or
\code{\link[ggpubr]{ggviolin}} (violin plot)
}
}
\description{
Use boxplots, violin plots or quantile plots (in form of parallel coordinate
plots of the quantiles) to compare the distribution of expression values for
each sample.
}
\examples{
data("se.gene")

## quantile plot
plot_distribution(se = se.gene,
                  method = "quantileplot")

## boxplot
plot_distribution(se = se.gene,
                  method = "boxplot")

## violinplot
library("ggpubr")
plot_distribution(se = se.gene,
                  method = "violinplot")
}
