% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_batch_effects.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap}
\usage{
plot_heatmap(matrix, type, title = NULL)
}
\arguments{
\item{matrix}{Matrix. Values to be plotted.}

\item{type}{Character. Type of heatmap ('pval' for P-values and
'r2' for absolute adjusted r^2).}

\item{title}{Character. Title of the plot.}
}
\value{
Heatmap as \code{\link[ComplexHeatmap]{Heatmap-class}} object
}
\description{
Plot heatmap with predefined color scheme based on type.
}
\examples{
data("se.gene")

res.pca = calculate_mds_pca(se = se.gene,
                            method = "pca")

col.test = c("Age.of.patient",
             "Body.surface.area",
             "Duration.of.psoriasis",
             "Induration",
             "Overall.erythema",
             "Scaling",
             "Sex",
             "scan.date")

res = check_batch_effects(se = se.gene,
                          res.pca = res.pca,
                          col.test = col.test)

# plot P-values
plot_heatmap(matrix = res$pval,
             type = "pval")

# plot absolute adjusted r^2
plot_heatmap(matrix = res$r2,
             type = "r2")
}
