% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds_pca.R
\name{plot_mds_pca}
\alias{plot_mds_pca}
\title{MDS or PCA plot}
\usage{
plot_mds_pca(res, var.color = NULL, palette = NULL, var.shape = NULL,
  shape.values = NULL, title = NULL, return.outliers = TRUE,
  factor = 5)
}
\arguments{
\item{res}{data.frame. Output of \code{\link{calculate_mds_pca}}.}

\item{var.color}{Character or integer vector. Variable used to determine
color. If NULL black color will be used for all samples.}

\item{palette}{Color palette to be used (default palette from
\code{\link[ggpubr]{get_palette}}).}

\item{var.shape}{Character or integer vector. Variable used to determine
shape. If NULL filled circles will be used for all samples.}

\item{shape.values}{Vector with symbols. Needs to provide a symbol for each
unique value of var.shape.}

\item{title}{Character. Title of the plot. If NULL title will be set based on
method.}

\item{return.outliers}{Logical. Return info about outlier samples.}

\item{factor}{Numeric. Parameter of the function
\code{\link[aplpack]{compute.bagplot}}. (default: 5)}
}
\value{
If return.outliers = TRUE, data.frame with identifiers of outlier
samples for each pairwise combination of components.
}
\description{
Plots components estimated with the function \code{\link{calculate_mds_pca}}.
Color and shape of each sample can be set based on different variables.
}
