% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds_pca.R
\name{plot_mds_pca_2d}
\alias{plot_mds_pca_2d}
\title{MDS or PCA plot (2D)}
\usage{
plot_mds_pca_2d(scores, dim, se, var.color = NULL, palette = NULL,
  var.shape = NULL)
}
\arguments{
\item{scores}{data.frame. Components of PCA or MDS analysis as estimated by
\code{\link{calculate_mds_pca}}.}

\item{dim}{Numeric vector (2). Numbers of components to be plotted.}

\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{var.color}{Character. Name of variable used to determine
color. If NULL black color will be used for all samples.}

\item{palette}{Color palette to be used (default palette used in
\code{\link[ggpubr]{ggscatter}}).}

\item{var.shape}{Character. Name of variable used to determine
shape. If NULL filled circles will be used for all samples.}
}
\description{
Plots two sepcified components estimated with the function
\code{\link{calculate_mds_pca}}. Color and shape of each sample can be set
based on different variables.
}
