% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds_pca.R
\name{prepare_var_for_plot}
\alias{prepare_var_for_plot}
\title{internal function used by plot_mds_pca}
\usage{
prepare_var_for_plot(se, var = NULL)
}
\description{
if var = NULL returns vector of 1s
}
\keyword{internal}
