% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_genes.R
\name{remove_genes}
\alias{remove_genes}
\title{Remove genes based on different criteria}
\usage{
remove_genes(se, assay, method, freq = 0.25, verbose = FALSE)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay to be used for
filtering.}

\item{method}{Method to determine genes to be removed: "constant", "missing",
"detection.pvalue", "zero".}

\item{freq}{Numeric. If more than freq*100 % of the samples fulfill criterion
the gene is removed (not used if method = "constant").}

\item{verbose}{Logical. Should number of removed genes be reported?}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with genes removed
}
\description{
Remove genes with no variability (method = "constant"), missing values
(method = "missing") or low expression (method = "detection.pvalue" or
method = "zero").
}
