% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_samples.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples based on different criteria}
\usage{
remove_samples(se, assay = 1, method, freq = 0.5, verbose = FALSE)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay to be used for
filtering.}

\item{method}{Method to determine samples to be removed: "missing",
"detection.pvalue", "zero".}

\item{freq}{Numeric. If more than freq*100 % of the genes fulfill criterion
the sample is removed.}

\item{verbose}{Logical. Should number of removed samples be reported?}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with samples removed
}
\description{
Remove samples with large frequency of missing values
(method = "missing") or low expression (method = "detection.pvalue" or
method = "zero").
}
