% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se.gene}
\alias{se.gene}
\title{Example data set (microarray, gene level)}
\format{
A
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}} with
12502 genes and 26 samples.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6710}
}
\usage{
se.gene
}
\description{
Microarray data from the GEO data set GSE6710 on the gene
level.
}
\keyword{datasets}
