% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se.probeset}
\alias{se.probeset}
\title{Example data set (microarray, probeset level)}
\format{
A
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}} with
22283 probesets and 5 samples.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6710}
}
\usage{
se.probeset
}
\description{
Microarray data from the GEO data set GSE6710 on the probeset
level restricted to the first five samples.
}
\keyword{datasets}
