% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_batch_effects.R
\name{check_batch_effects}
\alias{check_batch_effects}
\title{Check for batch effects}
\usage{
check_batch_effects(se, res.pca, col.test = NULL, title = NULL)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{res.pca}{List. Output of \code{\link{calculate_mds_pca}}.}

\item{col.test}{Character or integer vector. Column(s) in colData() with
phenotype information to be tested.}

\item{title}{Character. Title of the plot.}
}
\value{
List with the following components:
\itemize{
\item pval: Matrix with P-values between phenotype variables in rows and
components in columns
\item plot: Heatmap as \code{\link[ComplexHeatmap]{Heatmap-class}} object
}
}
\description{
Pairwise associations between each of the first three components of a MDS or
PCA analysis and defined phenotype variables is tested using F tests in a
linear model. P-values are visualized in a heatmap (called prince plot in
the R package swamp).
}
\examples{
data("se.gene")

res.pca = calculate_mds_pca(se = se.gene,
                            method = "pca")

col.test = c("Age.of.patient",
             "Body.surface.area",
             "Duration.of.psoriasis",
             "Induration",
             "Overall.erythema",
             "Scaling",
             "Sex",
             "scan.date")

check_batch_effects(se = se.gene,
                    res.pca = res.pca,
                    col.test = col.test)
}
