% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sex.R
\name{check_sex}
\alias{check_sex}
\title{Sex check}
\usage{
check_sex(se, assay = 1, symbol.column = "symbol",
  sex.column = "sex", col = c("red", "blue"), genes = c("DDX3Y",
  "EIF1AY", "KDM5D", "NLGN4Y", "RPS4Y1", "TXLNG2P", "UTY", "XIST"),
  title = NULL, verbose = FALSE)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay containing
expression data to be used for sex check.}

\item{symbol.column}{Character. Column in rowData containing gene symbols
(default: "symbol").}

\item{sex.column}{Character. Column in colData containing sex information.}

\item{col}{Character vector. Colours to be used for coloring sex in MDS plot
(default: red and blue).}

\item{genes}{Character vector. Symbols of sex specific genes to be used in
classification (default: genes collected from the literature, see details).}

\item{title}{Character. Title of the plot. If NULL title will be set to
"MDS (sex specific genes)".}

\item{verbose}{Logical. Should information about the genes be printed
(default: FALSE).}
}
\value{
List with the following components:
\itemize{
\item info: data.frame with information about samples with wrong sex or NULL
\item plot: MDS plot as returned by \code{plot_mds_pca_2d}
}
}
\description{
Predicts sex using k nearest neighbour classification based on expression
of sex specific genes.
}
\examples{
data("se.gene")

check_sex(se = se.gene,
          symbol.column = "Gene.symbol",
          sex.column = "Sex")
}
