% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicated_samples.R
\name{detect_duplicated_samples}
\alias{detect_duplicated_samples}
\title{Identification of duplicated samples}
\usage{
detect_duplicated_samples(se, assay = 1,
  corFinder.args = list(use.ComBat = FALSE), plot = FALSE, ...)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay containing
log transformed expression values.}

\item{corFinder.args}{List. Parameter of the function
\code{\link[doppelgangR]{doppelgangR}} (default: do not use
\code{\link[sva]{ComBat}} function for batch correction).}

\item{plot}{Logical. Should histogram of pairwise correlations be plotted?
(default: FALSE).}

\item{...}{Additional arguments passed to the
\code{\link[doppelgangR]{doppelgangR}} function.}
}
\value{
data.frame with information about duplicated samples (as generated by
summary function of \code{\link[doppelgangR]{DoppelGang-class}}) or NULL
}
\description{
Detects duplicated samples with extremely high correlation of expression
values using the function \code{\link[doppelgangR]{doppelgangR}}.
}
