% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize raw read counts of RNA-seq experiments}
\usage{
normalize_counts(se, assay.raw = 1, assay.norm = "counts.norm",
  method = "vst")
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay.raw}{Character or integer. Name or number of assay containing raw
read counts.}

\item{assay.norm}{Character. Name of new assay containing normalized
read counts.}

\item{method}{Method to use for normalizing: "vst" (default),
"rlog", "normTransform"}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with normalized read counts
}
\description{
Generates normalized read counts of RNA-seq data sets using several functions
provided in the Bioconductor package \pkg{DESeq2}.
}
