% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot distribution of expression values for each sample}
\usage{
plot_distribution(se, assay = 1, method = "quantileplot",
  return.outliers = TRUE, coef = 5)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay to be used for
plotting.}

\item{method}{Method to use for plotting: "quantileplot" (default),
"boxplot", "violinplot".}

\item{return.outliers}{Logical. Return info about outlier samples.}

\item{coef}{Numeric. Used in outlier definition (median +/- coef * IQR)}
}
\value{
If return.outliers = TRUE, data.frame with identifiers of outlier
samples
}
\description{
Use boxplots, violin plots or quantile plots (in form of parallel coordinate
plots of the quantiles) to compare the distribution of expression values for
each sample.
}
