% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds_pca.R
\name{plot_mds_pca}
\alias{plot_mds_pca}
\title{MDS or PCA plot}
\usage{
plot_mds_pca(res, se, var.color = NULL, palette = NULL,
  var.shape = NULL, shape.values = NULL, title = NULL, factor = 5,
  ellipse = FALSE, ellipse.type = "convex")
}
\arguments{
\item{res}{List. Output of \code{\link{calculate_mds_pca}}.}

\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{var.color}{Character or integer vector. Variable used to determine
color. If NULL black color will be used for all samples.}

\item{palette}{Color palette to be used (default palette from
\code{\link[ggpubr]{get_palette}}).}

\item{var.shape}{Character or integer vector. Variable used to determine
shape. If NULL filled circles will be used for all samples.}

\item{shape.values}{Vector with symbols. Needs to provide a symbol for each
unique value of var.shape.}

\item{title}{Character. Title of the plot. If NULL title will be set based
on method.}

\item{factor}{Numeric. Parameter of the function
\code{\link[aplpack]{compute.bagplot}}. (default: 5)}

\item{ellipse}{Logical. Should ellipses around points be drawn? (default:
FALSE).}

\item{ellipse.type}{Character. Type of ellipse as given in}
}
\value{
List with the following components:
\itemize{
\item info: data.frame with information about outlier samples for each
pairwise combination of component sor NULL
\item plot: Plot with three scatterplot as returned by
\code{\link[ggpubr]{ggarrange}}
}
}
\description{
Plots components estimated with the function \code{\link{calculate_mds_pca}}.
Color and shape of each sample can be set based on different variables.
}
\examples{
data("se.gene")

## PCA
res.pca = calculate_mds_pca(se = se.gene,
                            method = "pca")

## color code by group
plot_mds_pca(res = res.pca,
             se = se.gene,
             var.color = "group")
}
