% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds_pca.R
\name{plot_mds_pca_2d}
\alias{plot_mds_pca_2d}
\title{MDS or PCA plot (2D)}
\usage{
plot_mds_pca_2d(res, dim = c(1, 2), se, var.color = NULL,
  palette = NULL, var.shape = NULL, shape.values = NULL,
  title = NULL, factor = 5, ellipse = FALSE,
  ellipse.type = "convex")
}
\arguments{
\item{res}{List. Output of \code{\link{calculate_mds_pca}}.
\code{\link{calculate_mds_pca}}.}

\item{dim}{Numeric vector (2). Numbers of components to be plotted.}

\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{var.color}{Character. Name of variable used to determine
color. If NULL black color will be used for all samples.}

\item{palette}{Color palette to be used (default palette used in
\code{\link[ggpubr]{ggscatter}}).}

\item{var.shape}{Character. Name of variable used to determine
shape. If NULL filled circles will be used for all samples.}

\item{shape.values}{Vector with symbols. Needs to provide a symbol for each
unique value of var.shape.}

\item{title}{Character. Main title of plot.}

\item{factor}{Numeric. Parameter of the function
\code{\link[aplpack]{compute.bagplot}}. (default: 5)}

\item{ellipse}{Logical. Should ellipses around points be drawn? (default:
FALSE).}

\item{ellipse.type}{Character. Type of ellipse as given in
\code{\link[ggpubr]{ggscatter}} (default: "convex").}
}
\value{
List with the following components:
\itemize{
\item info: data.frame with information about outlier samples or NULL
\item plot: Scatterplot as returned by \code{\link[ggpubr]{ggscatter}}
}
}
\description{
Plots two sepcified components estimated with the function
\code{\link{calculate_mds_pca}}. Color and shape of each sample can be set
based on different variables.
}
\examples{
data("se.gene")

## PCA
res.pca = calculate_mds_pca(se = se.gene,
                            method = "pca")

## color code by group
plot_mds_pca_2d(res = res.pca,
                dim = c(1, 2),
                se = se.gene,
                var.color = "group")
}
